Minor refactor (BL-11581)
