+++ src/authentication/firebase/StyledFirebaseAuthCodeSplit.tsx
